Moved hooks relating to click listeners into hooks/click/
