--- jiu-jitsu-notes/src/hooks/useRandomTechnique.js
@@ -1,17 +1,0 @@
-import { useState } from "react";
-import { randomTechnique } from "../api/api";
-
-const useRandomTechnique = () => {
-  const [technique, setTechnique] = useState();
-  const [isLoading, setIsLoading] = useState(true);
-
-  const fetchRandomTechnique = async () => {
-    const newTechnique = await randomTechnique();
-    setTechnique(newTechnique);
-    setIsLoading(false);
-  };
-
-  return [fetchRandomTechnique, technique, isLoading];
-};
-
-export default useRandomTechnique;
